Stocks Data Visualization
Stock (also capital stock) of a corporation is all of the shares into which ownership of the corporation is divided. In American English, the shares are collectively known as "stock". A single share of the stock represents fractional ownership of the corporation in proportion to the total number of shares. 

Libraries requirement:
plotly==4.14.1
matplotlib==3.3.3
pandas==1.2.0
seaborn==0.11.1
numpy==1.19.3
scipy==1.5.4